Extract generic RoomTransport; NetworkManager becomes a thin adapter (#318 Step 2)
Increment 3 of the @usersfirst/room split. The game-agnostic transport is
now its own class/file; network-manager.js is reduced to a consumer of it.

- New js/net/room-transport.js (RoomTransport): room join-by-code, relay +
  P2P + fallback, reconnect/backoff, HEARTBEAT keepalive, media calls, and
  WebRTC stats. Sends/receives OPAQUE bytes — `_send(bytes)` out,
  `onMessage(bytes)` in — and contains zero bike/game knowledge (HEARTBEAT
  is its own reserved keepalive). This is the seam that lifts to a
  @usersfirst/room repo.
- network-manager.js: `class NetworkManager extends RoomTransport` (~95 lines)
  adds only the game layer — BikeCodec, the send* helpers, and the inbound
  dispatch wired via `this.onMessage = bytes => codec.dispatch(...)`. The
  pedal role-flip stays here (session state).

Done via inheritance specifically so the ENTIRE consumer surface (every
`this.net.*` field/method in lobby.js/game.js, including internals like
`_localMediaStream`) is preserved automatically — zero consumer edits, and
the wire format is unchanged.

Verified: room-transport has no game references; NetworkManager is a
RoomTransport; inherited transport methods + game send/receive all present;
inbound STATE decode and PEDAL role-flip route correctly through the
inherited _handleMessage → onMessage → codec.dispatch path.
